requirements.txt
beautifulsoup4==4.12.2
ordered-set==3.1.0
PyYAML==6.0
roman==4.1
selenium==4.10.0
thefuzz==0.19.0